requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
livereload==2.6.3
MarkupSafe==1.1.1
moviepy==1.0.3
numpy==1.18.1
opencv-python-headless==4.5.1.48
pandas==1.0.3
Pillow==7.1.2
pycodestyle==2.6.0
python-docx==0.8.10
python-pptx==0.6.18
six==1.15.0
toml==0.10.2
tornado==6.1
Werkzeug==1.0.1
